src/main/java/com/jingximall/jingximall/controllers/ProductController.java
package com.jingximall.jingximall.controllers;

import com.jingximall.jingximall.model.Product;
import com.jingximall.jingximall.repositories.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Sort;
import org.springframework.web.bind.annotation.*;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.PageRequest;

import java.util.List;

@RestController
@RequestMapping("/api/products")
public class ProductController {

    @Autowired
    ProductRepository productRepository;

    @GetMapping
    public List<Product> listProducts(
            @RequestParam(defaultValue = "0") int page,
            @RequestParam(defaultValue = "15") int size
    ) {
        Pageable pageable = PageRequest.of(page, size, Sort.by("id").ascending());
        var result =  productRepository.findAll(pageable);
        return result.getContent();
    }
}